Title: 컴퓨터와 함께하는 숫자 게임 2, Time: 279ms, Memory: 80MB, Status: Memory Limit Exceed - CodeTree

+++ 240129/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.cpp
@@ -6,7 +6,7 @@ using namespace std;
 
 int a;
 int b, c;
-vector<int> v;
+vector<long long> v;
 
 int func(int x){
     int left = 0;
